# packages/scripts/src/etl/jenny-v3/jenny-v3-openapi.yaml
openapi: 3.0.0
info:
  title: Jenny v3 API
  version: 3.0.0
  description: Evidence-based AI with vitals-first architecture
servers:
  - url: http://localhost:8787
    description: Local development
paths:
...
  /students/{student_id}/vitals:
    get:
      summary: Get student vitals (facts from PostgreSQL)
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            example: huda-2025
      responses:
        '200':
          description: Student vitals
          content:
            application/json:
              schema:
                type: object
                properties:
                  student_id:
                    type: string
                  facts:
                    type: array
                    items:
                      type: object
                      properties:
                        kind:
                          type: string
                          example: sat_total_score
                        value:
                          type: string
                          example: "1530"
                        fact_date:
                          type: string
                          format: date
                        confidence:
                          type: string
                          enum: [high, medium, low]
                        source_id:
                          type: string
  /students/{student_id}/lifecycle:
    get:
      summary: Get student lifecycle items
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lifecycle items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    item_id:
                      type: string
                    domain:
                      type: string
                      enum: [application, award, test, essay, recommender, ec_portfolio, aid_css_fafsa, ops_policy]
                    status:
                      type: string
                      enum: [planned, in_progress, submitted, outcome, archived]
                    school:
                      type: string
                      nullable: true
                    submitted_at:
                      type: string
                      format: date-time
                      nullable: true
                    outcome_date:
                      type: string
                      format: date-time
                      nullable: true
...